+++ src/models/command/blockAutomaticCommand.ts
@@ -2,24 +2,69 @@ import { CommandType } from './commandType';
 import { CommandSource } from './commandSource';
 import { CollisionSolving } from '../collisionSolving';
 import { BaseCommand } from './baseCommand';
+import { SettingsService } from '../../server';
+import { DeviceSettings } from '../deviceSettings';
 
 export class BlockAutomaticCommand extends BaseCommand {
   /** @inheritDoc */
   public _commandType: CommandType = CommandType.BlockAutomaticCommand;
+  /**
+   * Whether the device should revert to automatic afterward.
+   */
+  public readonly revertToAutomaticAtBlockLift: boolean;
+
+  /**
+   * The duration in milliseconds for the automatic actions to be disabled.
+   */
+  public readonly durationMS: number;
+
+  /**
+   * The action to take if a block is already active.
+   */
+  public readonly onCollideAction: CollisionSolving;
 
   /**
    * Command to disable automatic actions for a specific duration.
    * @param source - The source of the command.
-   * @param durationMS - The duration in milliseconds for the automatic actions to be disabled.
+   * @param durationMS - The duration in milliseconds for the automatic actions to be disabled --> If unset the default value will be used {@link SettingsService.settings.blockAutomaticHandlerDefaults.blockAutomaticDurationMS}
    * @param reason - You can provide an individual reason here for debugging purpose.
-   * @param onCollideAction - The action to take if a block is already active. --> Default: overrideIfGreater
+   * @param onCollideAction - The action to take if a block is already active. --> Default: {@link SettingsService.settings.blockAutomaticHandlerDefaults.defaultCollisionSolving}
+   * @param revertToAutomaticAtBlockLift - Whether the device should revert to automatic afterward. --> Default: {@link SettingsService.settings.blockAutomaticHandlerDefaults.revertToAutomaticAtBlockLift}
    */
   public constructor(
     source: CommandSource | BaseCommand,
-    public durationMS: number,
+    durationMS?: number,
     reason: string = '',
-    public onCollideAction: CollisionSolving = CollisionSolving.overrideIfGreater,
+    onCollideAction?: CollisionSolving,
+    revertToAutomaticAtBlockLift?: boolean,
   ) {
     super(source, reason);
+    this.durationMS =
+      durationMS ?? SettingsService.settings?.blockAutomaticHandlerDefaults?.blockAutomaticDurationMS ?? 30 * 60 * 1000;
+    this.revertToAutomaticAtBlockLift =
+      revertToAutomaticAtBlockLift ??
+      SettingsService.settings?.blockAutomaticHandlerDefaults?.revertToAutomaticAtBlockLift ??
+      true;
+    this.onCollideAction =
+      onCollideAction ??
+      SettingsService.settings?.blockAutomaticHandlerDefaults?.defaultCollisionSolving ??
+      CollisionSolving.overrideIfGreater;
+  }
+
+  public get logMessage(): string {
+    return `BlockAutomatic for ${this.durationMS}ms with onCollideAction ${this.onCollideAction} for reason: ${this.reasonTrace}`;
+  }
+
+  public static fromDeviceSettings(c: BaseCommand, deviceSettings: DeviceSettings): BlockAutomaticCommand | null {
+    if (deviceSettings.blockAutomaticSettings?.dontBlockAutomaticIfNotProvided) {
+      return null;
+    }
+    return new BlockAutomaticCommand(
+      c,
+      deviceSettings.blockAutomaticSettings?.blockAutomaticDurationMS,
+      '',
+      deviceSettings.blockAutomaticSettings?.defaultCollisionSolving,
+      deviceSettings.blockAutomaticSettings?.revertToAutomaticAtBlockLift,
+    );
   }
 }