+++ src/models/command/blockAutomaticUntilCommand.ts
@@ -2,24 +2,44 @@ import { CommandType } from './commandType';
 import { CommandSource } from './commandSource';
 import { CollisionSolving } from '../collisionSolving';
 import { BaseCommand } from './baseCommand';
+import { SettingsService } from '../../server';
 
 export class BlockAutomaticUntilCommand extends BaseCommand {
   /** @inheritDoc */
   public _commandType: CommandType = CommandType.BlockAutomaticUntilCommand;
 
   /**
+   * The action to take if a block is already active.
+   */
+  public readonly onCollideAction: CollisionSolving;
+  /**
+   * Whether the device should revert to automatic afterward.
+   */
+  public readonly revertToAutomaticAtBlockLift: boolean;
+
+  /**
    * Command to disable automatic actions until a specific date.
    * @param source - The source of the command.
    * @param targetDate - The date until the automatic actions will be disabled.
    * @param reason - You can provide an individual reason here for debugging purpose.
-   * @param onCollideAction - The action to take if a block is already active. --> Default: overrideIfGreater
+   * @param onCollideAction - The action to take if a block is already active. --> Default: {@link SettingsService.settings.blockAutomaticHandlerDefaults.defaultCollisionSolving}
+   * @param revertToAutomaticAtBlockLift - Whether the device should revert to automatic afterward. --> Default: {@link SettingsService.settings.blockAutomaticHandlerDefaults.revertToAutomaticAtBlockLift}
    */
   public constructor(
     source: CommandSource | BaseCommand,
     public targetDate: Date,
     reason: string = '',
-    public onCollideAction: CollisionSolving = CollisionSolving.overrideIfGreater,
+    onCollideAction?: CollisionSolving,
+    revertToAutomaticAtBlockLift?: boolean,
   ) {
     super(source, reason);
+    this.onCollideAction =
+      onCollideAction ??
+      SettingsService.settings?.blockAutomaticHandlerDefaults?.defaultCollisionSolving ??
+      CollisionSolving.overrideIfGreater;
+    this.revertToAutomaticAtBlockLift =
+      revertToAutomaticAtBlockLift ??
+      SettingsService.settings?.blockAutomaticHandlerDefaults?.revertToAutomaticAtBlockLift ??
+      true;
   }
 }