+++ src/models/command/wledSetLightCommand.ts
@@ -5,6 +5,7 @@ import { DimmerSetLightCommand } from './dimmerSetLightCommand';
 import { WledSettings } from '../deviceSettings';
 import { LampSetTimeBasedCommand } from './lampSetTimeBasedCommand';
 import { TimeOfDay } from '../timeCallback';
+import { BlockAutomaticCommand } from './blockAutomaticCommand';
 
 export class WledSetLightCommand extends DimmerSetLightCommand {
   /** @inheritDoc */
@@ -15,7 +16,8 @@ export class WledSetLightCommand extends DimmerSetLightCommand {
    * @param source - The source of the command
    * @param on - The desired state of the light
    * @param reason - You can provide a reason for clarification
-   * @param timeout - The duration in milliseconds this should block automatic changes --> Reverse to automatic after this time
+   * @param disableAutomatic - If provided, the device will remain in the desired state for the given disable action.
+   * If unset the default value will be used: {@link SettingsService.settings.blockAutomaticHandlerDefaults}
    * @param brightness - The brightness of the light
    * @param transitionTime - The time in milliseconds the transition should take
    * @param preset - The preset to use
@@ -24,16 +26,16 @@ export class WledSetLightCommand extends DimmerSetLightCommand {
     source: CommandSource | BaseCommand,
     on: boolean,
     reason: string = '',
-    timeout: number = -1,
+    disableAutomatic?: BlockAutomaticCommand | null,
     brightness: number = -1,
     transitionTime: number = -1,
     public preset?: number,
   ) {
-    super(source, on, reason, timeout, brightness, transitionTime);
+    super(source, on, reason, disableAutomatic, brightness, transitionTime);
   }
 
   public override get logMessage(): string {
-    return `Dimmer setLight to ${this.on} with Brightness ${this.brightness}, timeout ${this.timeout} and preset ${this.preset} for reason: ${this.reasonTrace}`;
+    return `Dimmer setLight to ${this.on} with Brightness ${this.brightness}, disabelAutomatic ${this.disableAutomaticCommand?.logMessage} and preset ${this.preset} for reason: ${this.reasonTrace}`;
   }
 
   public static override byTimeBased(settings: WledSettings, c: LampSetTimeBasedCommand): WledSetLightCommand {
@@ -43,7 +45,7 @@ export class WledSetLightCommand extends DimmerSetLightCommand {
           c,
           settings.dayOn,
           '',
-          c.timeout,
+          c.disableAutomaticCommand,
           settings.dayBrightness,
           undefined,
           settings.dayPreset,
@@ -53,7 +55,7 @@ export class WledSetLightCommand extends DimmerSetLightCommand {
           c,
           settings.dawnOn,
           '',
-          c.timeout,
+          c.disableAutomaticCommand,
           settings.dawnBrightness,
           undefined,
           settings.dawnPreset,
@@ -63,7 +65,7 @@ export class WledSetLightCommand extends DimmerSetLightCommand {
           c,
           settings.duskOn,
           '',
-          c.timeout,
+          c.disableAutomaticCommand,
           settings.duskBrightness,
           undefined,
           settings.duskPreset,
@@ -73,7 +75,7 @@ export class WledSetLightCommand extends DimmerSetLightCommand {
           c,
           settings.nightOn,
           '',
-          c.timeout,
+          c.disableAutomaticCommand,
           settings.nightBrightness,
           undefined,
           settings.nightPreset,